Add real build to CI
Adds the build-batch job to CI which actually builds the entire
snapshot. It does this by dividing the snapshot into batch groups
per --batch, then having a different job build each group via
github action's matrix strategy and new clc-stackage --batch-index
param.

This allows us to build the entire snapshot such that overall CI time
is reasonable i.e. ~ 1 hour.

+++ .github/scripts/dry_run.sh
@@ -1,15 +1,19 @@
 set -e
 
-echo "*** Updating cabal ***"
+if [[ -f ./bin/clc-stackage ]]; then
+  echo "*** ./bin/clc-stackage exists, not re-installing ***"
+  chmod a+x ./bin/clc-stackage
+else
+  echo "*** Updating cabal ***"
+  cabal update
 
-cabal update
+  # --overwrite-policy=always and deleting output/ are unnecessary for CI since
+  # this script will only be run one time, but it's helpful when we are
+  # testing the script locally.
 
-echo "*** Installing clc-stackage ***"
-
-# --overwrite-policy=always and deleting output/ are unnecessary for CI since
-# this script will only be run one time, but it's helpful when we are
-# testing the script locally.
-cabal install exe:clc-stackage --installdir=./bin --overwrite-policy=always
+  echo "*** Installing clc-stackage ***"
+  cabal install exe:clc-stackage --installdir=./bin --overwrite-policy=always
+fi
 
 if [[ -d output ]]; then
   rm -r output
@@ -18,7 +22,7 @@ fi
 echo "*** Building all with --dry-run ***"
 
 set +e
-./bin/clc-stackage --batch 100 --cabal-options="--dry-run"
+./bin/clc-stackage --batch 200 --cabal-options="--dry-run"
 
 ec=$?
